requirements.txt
fastapi
"tortoise-orm[asyncpg]"
uvicorn
aerich
python-dotenv
